+++ src/runtime/threads/analyzer.c
@@ -12,8 +12,8 @@
 void *analyzer_func(void *vargp) {
   (void)vargp;
 
-  int cpus = get_cpu_count();
-  int buffer_size = (sizeof(char) * sizeof(double)) * cpus;
+  unsigned int cpus = get_cpu_count();
+  size_t buffer_size = (size_t)((sizeof(char) * sizeof(double)) * cpus);
   char *buffer = malloc(buffer_size);
   char *msg = malloc(sizeof(char) * 30);
 
@@ -21,9 +21,9 @@ void *analyzer_func(void *vargp) {
     pt_set_alive(pt_mutex_analyzer_alive, true);
 
     char *old = pt_queue_dequeue(pt_queue_reader_analyzer,
-                                 (PROC_LINE_LENGTH + 1) * cpus);
+                                 (size_t)(PROC_LINE_LENGTH + 1) * cpus);
     char *new = pt_queue_dequeue(pt_queue_reader_analyzer,
-                                 (PROC_LINE_LENGTH + 1) * cpus);
+                                 (size_t)(PROC_LINE_LENGTH + 1) * cpus);
 
     msg = "Analisys in progress...";
     pt_queue_enqueue(pt_queue_logger, msg, strlen(msg) + 1);
@@ -43,7 +43,6 @@ void *analyzer_func(void *vargp) {
     msg = "Analisys done.";
     pt_queue_enqueue(pt_queue_logger, msg, strlen(msg) + 1);
   }
-
 
   free(buffer);
   free(msg);
